requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
cffi==1.14.4
chardet==3.0.4
discord.py==1.5.1
environs==8.0.0
idna==2.10
marshmallow==3.7.1
multidict==4.7.6
pycparser==2.20
PyNaCl==1.3.0
python-dotenv==0.14.0
six==1.15.0
yarl==1.5.1
